backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.9",
    "msoffcrypto-tool>=6.0.0",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "unstructured>=0.21.5",
    "uvicorn>=0.41.0",
]
